refactor(config): configurations
This cleans up config & entry point of application making it easier to
manage & maintain. Moving context out of the server entry point allows
the context to be expanded to handle further usecases that will be
testable independently as it is a function on its own with it's own
logic. This way the application can be extended further to handle even
more usecases.